javadoc and cleanup
