+++ indexing-service/src/main/java/org/apache/druid/indexing/overlord/http/OverlordResource.java
@@ -62,10 +62,10 @@ import org.apache.druid.java.util.common.Intervals;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.metadata.EntryExistsException;
-import org.apache.druid.metadata.MetadataStorageActionHandler.ActiveTaskLookup;
-import org.apache.druid.metadata.MetadataStorageActionHandler.CompleteTaskLookup;
-import org.apache.druid.metadata.MetadataStorageActionHandler.TaskLookup;
-import org.apache.druid.metadata.MetadataStorageActionHandler.TaskLookupType;
+import org.apache.druid.metadata.TaskLookup;
+import org.apache.druid.metadata.TaskLookup.ActiveTaskLookup;
+import org.apache.druid.metadata.TaskLookup.CompleteTaskLookup;
+import org.apache.druid.metadata.TaskLookup.TaskLookupType;
 import org.apache.druid.server.http.HttpMediaType;
 import org.apache.druid.server.http.security.ConfigResourceFilter;
 import org.apache.druid.server.http.security.DatasourceResourceFilter;
