+++ server/src/test/java/org/apache/druid/metadata/SQLMetadataStorageActionHandlerTest.java
@@ -31,8 +31,8 @@ import org.apache.druid.java.util.common.DateTimes;
 import org.apache.druid.java.util.common.Pair;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.jackson.JacksonUtils;
-import org.apache.druid.metadata.MetadataStorageActionHandler.ActiveTaskLookup;
-import org.apache.druid.metadata.MetadataStorageActionHandler.CompleteTaskLookup;
+import org.apache.druid.metadata.TaskLookup.ActiveTaskLookup;
+import org.apache.druid.metadata.TaskLookup.CompleteTaskLookup;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
